+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## v4.62.0 Fixes
 
 - `[Homepage]` Fixed instability of the visual tests. ([#6179](https://github.com/infor-design/enterprise/issues/6179))
+- `[Modal]` Added closeButton method that defaults to force close even if other subcomponents are still open. ([#6048](https://github.com/infor-design/enterprise/issues/6048))
 - `[Widget]` Fix on drag image including the overflow area. ([NG#1216](https://github.com/infor-design/enterprise-ng/issues/1216))
 
 ## v4.61.0 Features